refactor: move file_processing modules
