+++ tests/file_processing/test_format_detector.py
@@ -3,8 +3,11 @@ from pathlib import Path
 import pandas as pd
 import pytest
 
-from file_processing.format_detector import FormatDetector, UnsupportedFormatError
-from file_processing.readers import (
+from yosai_intel_dashboard.src.file_processing.format_detector import (
+    FormatDetector,
+    UnsupportedFormatError,
+)
+from yosai_intel_dashboard.src.file_processing.readers import (
     ArchiveReader,
     CSVReader,
     ExcelReader,
