+++ tests/test_file_processor.py
@@ -7,7 +7,7 @@ import io
 import pandas as pd
 import pytest
 
-from file_processing import create_file_preview
+from yosai_intel_dashboard.src.file_processing import create_file_preview
 from services.data_processing.file_processor import (
     UnicodeFileProcessor,
     process_uploaded_file,
