+++ tests/test_max_display_rows.py
@@ -9,7 +9,7 @@ from tests.fake_configuration import FakeConfiguration
 
 fake_cfg = FakeConfiguration()
 from config.constants import MAX_DISPLAY_ROWS
-from file_processing import create_file_preview
+from yosai_intel_dashboard.src.file_processing import create_file_preview
 
 
 def test_dynamic_config_default_display_rows():
