+++ yosai_intel_dashboard/src/services/upload/utils/__init__.py
@@ -1,6 +1,6 @@
 """Upload utility helpers."""
 
-from file_processing import create_file_preview
+from yosai_intel_dashboard.src.file_processing import create_file_preview
 
 from .file_parser import UnicodeFileProcessor
